+++ app/models/daily_cache.rb
@@ -171,7 +171,7 @@ class DailyCache
 
       end
 
-      @metrics << { :name => sensor.name, :data => count, :range => time_range }
+      @metrics << { :name => sensor.sensor_name, :data => count, :range => time_range }
     end
 
     @metrics