+++ config/application.rb
@@ -12,17 +12,24 @@ Bundler.require(:default, Rails.env) if defined?(Bundler)
 
 module Snorby
 
+  # Check Ruby Version
+  unless RUBY_VERSION.match(/^1.9/)
+    puts "Snorby requires Ruby version 1.9.x"
+    puts "We suggest using Ruby Version Manager (RVM) https://rvm.io/ to install the newest release"
+    exit 1
+  end
+
   # Check For snorby_config.yml
   unless File.exists?("config/snorby_config.yml")
     puts "Snorby Configuration Error"
-    puts "* Please EDIT and rename config/snorby_config.example.yml to config/snorby_config.yml"
+    puts "* Please EDIT and rename config/snorby_config.yml.example to config/snorby_config.yml"
     exit 1
   end
   
   # Check For database.yml
   unless File.exists?("config/database.yml")
     puts "Snorby Configuration Error"
-    puts "* Please EDIT and rename config/database.example.yml to config/database.yml"
+    puts "* Please EDIT and rename config/database.yml.example to config/database.yml"
     exit 1
   end
 
@@ -36,7 +43,12 @@ module Snorby
   end
 
   class Application < Rails::Application
-        
+
+    config.threadsafe!
+    config.dependency_loading = true
+    # Custom directories with classes and modules you want to be autoloadable.
+    config.autoload_paths += %W(#{config.root}/lib)
+
     PDFKit.configure do |config|
       config.wkhtmltopdf = Snorby::CONFIG[:wkhtmltopdf]
       config.default_options = {
@@ -44,14 +56,11 @@ module Snorby
           :print_media_type => true
         }
     end
-    
+
     # Settings in config/environments/* take precedence over those specified here.
     # Application configuration should go into files in config/initializers
     # -- all .rb files in that directory are automatically loaded.
 
-    # Custom directories with classes and modules you want to be autoloadable.
-    config.autoload_paths += %W(#{config.root}/lib)
-    
     # Only load the plugins named here, in the order given (default is alphabetical).
     # :all can be used as a placeholder for all plugins not explicitly named.
     # config.plugins = [ :exception_notification, :ssl_requirement, :all ]
@@ -84,7 +93,10 @@ module Snorby
     # Configure sensitive parameters which will be filtered from the log file.
     config.filter_parameters += [:password]
 
-    # SSL
-    # config.force_ssl = true
+    # Enable SSL if it was enabled in the configuration
+    if CONFIG.has_key?(:ssl) && CONFIG[:ssl]
+      config.force_ssl = true
+    end
   end
+
 end