+++ config/environment.rb
@@ -1,5 +1,3 @@
 # Load the rails application
 require File.expand_path('../application', __FILE__)
-
-# Initialize the rails application
 Snorby::Application.initialize!