+++ config/environments/development.rb
@@ -14,6 +14,9 @@ Snorby::Application.configure do
   #config.action_view.debug_rjs             = true
   config.action_controller.perform_caching = false
 
+  # Send mail to letter opener gem to open generated mails in a new window
+  config.action_mailer.delivery_method = :letter_opener
+
   # Don't care if the mailer can't send
   config.action_mailer.raise_delivery_errors = true
 
@@ -23,5 +26,7 @@ Snorby::Application.configure do
   # Only use best-standards-support built into browsers
   config.action_dispatch.best_standards_support = :builtin
   
+  # config.threadsafe!
+  # config.dependency_loading = true
 end