+++ lib/snorby/paperclip_processors/cropper.rb
@@ -1,37 +1,19 @@
-# Snorby - All About Simplicity.
-# 
-# Copyright (c) 2010 Dustin Willis Webber (dustin.webber at gmail.com)
-# 
-# This program is free software; you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation; either version 2 of the License, or
-# (at your option) any later version.
-# 
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-# 
-# You should have received a copy of the GNU General Public License
-# along with this program; if not, write to the Free Software
-# Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
+# module Paperclip
+  # class Cropper < Thumbnail
+    # def transformation_command
+      # if crop_command
+        # crop_command + super.sub(/ -crop \S+/, '')
+      # else
+        # super
+      # end
+    # end
 
-module Paperclip
-  class Cropper < Thumbnail
-    def transformation_command
-      if crop_command
-        crop_command + super.sub(/ -crop \S+/, '')
-      else
-        super
-      end
-    end
+    # def crop_command
+      # target = @attachment.instance
+      # if target.cropping?
+        # " -crop '#{target.crop_w.to_i}x#{target.crop_h.to_i}+#{target.crop_x.to_i}+#{target.crop_y.to_i}' "
+      # end
+    # end
 
-    def crop_command
-      target = @attachment.instance
-      if target.cropping?
-        " -crop '#{target.crop_w.to_i}x#{target.crop_h.to_i}+#{target.crop_x.to_i}+#{target.crop_y.to_i}' "
-      end
-    end
-
-  end
-end
+  # end
+# end