+++ ChangeLog.md
@@ -1,3 +1,11 @@
+# Snorby 2.5.2
+
+* License Changed to GPLv3 - Please review the License file.
+* You can now delete sensors.
+* Timezone added to all DateTime
+
+* Note: Make sure to run bundle exec rake snorby:update
+
 # Snorby 2.5.1
 
   * Fixed Firefox JS/search issues
@@ -8,6 +16,8 @@
   * Fixed a timestamp/sql format issue with older versions of mysql
   * Fixed bug when classification is zero (unclassified)
   * Fixed login image copyright date
+
+  * Note: Make sure to run bundle exec rake snorby:update
 
 # Snorby 2.5.0